sam: use ulong as type of dev consistently
Otherwise io.c:39 compares int != ulong,
causing spurious file modification warnings
when dev doesn't fit in an int.

R=r
http://codereview.appspot.com/1917045

+++ src/cmd/sam/sam.h
@@ -160,7 +160,7 @@ struct File
 	String		name;		/* name of associated file */
 	uvlong		qidpath;	/* of file when read */
 	uint		mtime;		/* of file when read */
-	int		dev;		/* of file when read */
+	ulong	dev;		/* of file when read */
 	int		unread;		/* file has not been read from disk */
 
 	long		seq;		/* if seq==0, File acts like Buffer */
